requirements.txt
anyascii==0.3.1
asgiref==3.6.0
beautifulsoup4==4.9.3
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.0.10
django-filter==21.1
django-modelcluster==5.3
django-taggit==2.1.0
django-treebeard==4.6.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
l18n==2021.3
openpyxl==3.1.2
pi==0.1.2
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tablib==3.3.0
telepath==0.3
urllib3==1.26.15
wagtail==2.16.3
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.9
xlwt==1.3.0
